+++ packages/wallets/ledger/src/clients/thorchain/lib.ts
@@ -14,8 +14,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  ******************************************************************************* */
+import { bech32 } from '@scure/base';
 import { ErrorCode } from '@thorswap-lib/types';
-import { bech32 } from 'bech32';
 import crypto from 'crypto';
 import Ripemd160 from 'ripemd160';
 
@@ -64,7 +64,12 @@ export class THORChainApp {
     }
     const hashSha256 = crypto.createHash('sha256').update(pk).digest();
     const hashRip = new Ripemd160().update(hashSha256).digest();
-    return bech32.encode(hrp, bech32.toWords(hashRip));
+    // @ts-ignore
+    const encode = bech32.encode || bech32.bech32.encode;
+    // @ts-ignore
+    const toWords = bech32.toWords || bech32.bech32.toWords;
+
+    return encode(hrp, toWords(hashRip));
   }
 
   async serializePath(path: string) {