+++ packages/wallets/walletconnect/src/walletconnect.ts
@@ -1,5 +1,6 @@
 import type { StdSignDoc } from '@cosmjs/amino';
-import type { TxBodyEncodeObject } from '@cosmjs/proto-signing';
+import { Int53 } from '@cosmjs/math';
+import { encodePubkey, makeAuthInfoBytes, type TxBodyEncodeObject } from '@cosmjs/proto-signing';
 import type { DepositParam } from '@thorswap-lib/toolbox-cosmos';
 import type { WalletTxParams } from '@thorswap-lib/types';
 import { ApiUrl, Chain, ChainId, WalletOption } from '@thorswap-lib/types';
@@ -156,11 +157,9 @@ const getToolbox = async ({
           value: sendCoinsMessage,
         };
 
-        const { encodePubkey, makeAuthInfoBytes } = await import('@cosmjs/proto-signing');
         const { makeSignDoc } = await import('@cosmjs/amino');
         const { fromBase64 } = await import('@cosmjs/encoding');
         const { StargateClient } = await import('@cosmjs/stargate');
-        const { Int53 } = await import('@cosmjs/math');
 
         const signDoc = makeSignDoc(
           [msg],