+++ packages/toolboxes/toolbox-cosmos/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @thorswap-lib/toolbox-cosmos
 
+## 1.0.0-development.155
+
+### Patch Changes
+
+- Maybe fix import'
+
 ## 1.0.0-development.154
 
 ### Patch Changes