+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @thorswap-lib/keplr
 
+## 1.0.0-development.50
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thorswap-lib/toolbox-cosmos@1.0.0-development.155
+
 ## 1.0.0-development.49
 
 ### Patch Changes